LayoutTests/fast/forms/input-value-sanitization-expected.txt

Tests for value sanitization algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Number:
PASS input.value is "65536"
PASS input.value = "256"; input.value is "256"
PASS input.value = ""; input.value is ""

Range:
PASS input.value is "50"
PASS successfullyParsed is true

TEST COMPLETE
